requirements.txt
discord.py>=1.2.0
aiohttp<3.6.0
tinydb>=3.15.0
